refactor: splitting into different files for greater eligibility
